refactor: update logger path
